🚚  zustand sales code split 하기 위해 폴더 생성 및 이동
