+++ src/components/sales/calendar/header/Header.tsx
@@ -1,4 +1,4 @@
-import useSalesStore from '@/shared/store/sales';
+import useSalesStore from '@/shared/store/sales/sales';
 import clsx from 'clsx';
 import HeaderController from './calendarController/HeaderController';
 import HeaderDate from './headerDate/HeaderDate';
