+++ src/components/sales/calendar/header/calendarController/HeaderController.tsx
@@ -1,6 +1,6 @@
 import Button from '@/components/common/Button';
 import { useCalendar } from '@/hooks/sales/useCalendar';
-import useSalesStore from '@/shared/store/sales';
+import useSalesStore from '@/shared/store/sales/sales';
 import clsx from 'clsx';
 import styles from './styles/headerController.module.css';
 import ArrowLeft from '/public/icons/calendar-arrow-left.svg';
