+++ src/components/sales/calendar/header/headerDate/HeaderDate.tsx
@@ -1,4 +1,4 @@
-import useSalesStore from '@/shared/store/sales';
+import useSalesStore from '@/shared/store/sales/sales';
 import clsx from 'clsx';
 import styles from './styles/headerDate.module.css';
 const HeaderDate = () => {
