+++ src/components/sales/status/SalesStatus.tsx
@@ -1,5 +1,5 @@
 import { getDaySales } from '@/server/api/supabase/sales';
-import useSalesStore from '@/shared/store/sales';
+import useSalesStore from '@/shared/store/sales/sales';
 import useAuthState from '@/shared/store/session';
 import { Tables } from '@/types/supabase';
 import moment from 'moment';
